indicate spring/winter solstices
This is done with a threshold on solar altitude, for the day being
shown.  I picked a threshold that detected the spring and winter
solstices for years 2019 and 2020, so I hope they will work for other
years. Since this is done with a threshold, the scheme will be brittle
to the time between calculations, so do *not* change the seq() call that
establishes that.  I found that even a slightly larger threshold will
detect two days for winter solstice in 2019.